backend/pyproject.toml
[project]
name = "backend"
version = "2.0.0"
description = "Battery Cell Dashboard API with database persistence"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "impedance>=1.7.1",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "cloudinary>=1.40.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
